Added optional sync-RN step to Rollup build script
