+++ scripts/rollup/build.js
@@ -18,6 +18,7 @@ const Bundles = require('./bundles');
 const propertyMangleWhitelist = require('./mangle').propertyMangleWhitelist;
 const sizes = require('./plugins/sizes-plugin');
 const Stats = require('./stats');
+const syncReactNative = require('./sync').syncReactNative;
 const Packaging = require('./packaging');
 const Header = require('./header');
 
@@ -37,6 +38,7 @@ const requestedBundleTypes = (argv.type || '')
 const requestedBundleNames = (argv._[0] || '')
   .split(',')
   .map(type => type.toLowerCase());
+const syncRN = argv['sync-rn'];
 
 // used for when we property mangle with uglify/gcc
 const mangleRegex = new RegExp(
@@ -471,6 +473,9 @@ rimraf('build', () => {
       () => createBundle(bundle, RN_DEV),
       () => createBundle(bundle, RN_PROD)
     );
+  }
+  if (syncRN) {
+    tasks.push(() => syncReactNative(join('build', 'react-native'), syncRN));
   }
   // rather than run concurently, opt to run them serially
   // this helps improve console/warning/error output
